s390/boot: Introduce leave_decompressor label
Introduce a global label that is very close to the jump from the
decompressor to the decompressed kernel. Put its address into
svc_old_psw, from where GDB can take it without loading decompressor
symbols. It should be available throughout the entire decompressor
execution, because it's placed there statically, and nothing in the
decompressor uses the SVC instruction.

Signed-off-by: Ilya Leoshkevich <iii@linux.ibm.com>