perf test java symbol: Add PID namespace variant
Add a PID namespace variant of test_java_symbol.sh. Compared to the
existing test, we must jump through two extra hoops here.

First, we need to check whether unshare is there an is usable, which
may not be the case either due to kernel config or permissions.

Second, we need to keep jshell running after perf record exits in order
to reproduce a real issue with this setup. This means we have to attach
perf after starting jshell, use ctl-fifo and ack-fifo, and work around
perf's inability to attach to existing children.

Signed-off-by: Ilya Leoshkevich <iii@linux.ibm.com>